refactor(key-wallet): split ManagedCoreAccount into funds + keys variants
Move the existing struct out of `managed_account/mod.rs` into a dedicated
`managed_core_funds_account.rs` and rename it to `ManagedCoreFundsAccount`
to make room for a sibling `ManagedCoreKeysAccount` that omits funds-tracking
state (`balance`, `utxos`, `spent_outpoints`). The keys variant is intended
for accounts that derive special-purpose keys (identity registration, asset
locks, masternode provider keys) but is not yet wired into
`ManagedAccountCollection` — that migration follows in a subsequent PR.

This change is purely structural: `ManagedAccountCollection` continues to
store `ManagedCoreFundsAccount` for every account type and behavior is
unchanged. The C ABI is preserved — `FFIManagedCoreAccount` and the
`managed_core_account_*` C functions keep their original names and now wrap
the renamed Rust struct internally.
