+++ key-wallet-ffi/src/utxo_tests.rs
@@ -175,7 +175,7 @@ mod utxo_tests {
         use dashcore::blockdata::script::ScriptBuf;
         use dashcore::{Address, OutPoint, TxOut, Txid};
         use key_wallet::account::account_type::StandardAccountType;
-        use key_wallet::managed_account::ManagedCoreAccount;
+        use key_wallet::managed_account::ManagedCoreFundsAccount;
         use key_wallet::utxo::Utxo;
         use key_wallet::wallet::managed_wallet_info::ManagedWalletInfo;
         use key_wallet::Network;
@@ -189,7 +189,7 @@ mod utxo_tests {
         let mut managed_info = ManagedWalletInfo::new(Network::Testnet, [1u8; 32]);
 
         // Create a BIP44 account with UTXOs
-        let mut bip44_account = ManagedCoreAccount::new(
+        let mut bip44_account = ManagedCoreFundsAccount::new(
             ManagedAccountType::Standard {
                 index: 0,
                 standard_account_type: StandardAccountType::BIP44Account,
@@ -280,7 +280,7 @@ mod utxo_tests {
     fn test_managed_wallet_get_utxos_multiple_accounts() {
         use crate::managed_wallet::FFIManagedWalletInfo;
         use key_wallet::account::account_type::StandardAccountType;
-        use key_wallet::managed_account::ManagedCoreAccount;
+        use key_wallet::managed_account::ManagedCoreFundsAccount;
         use key_wallet::wallet::managed_wallet_info::ManagedWalletInfo;
         use key_wallet::Network;
 
@@ -292,7 +292,7 @@ mod utxo_tests {
         let mut managed_info = ManagedWalletInfo::new(Network::Testnet, [2u8; 32]);
 
         // Create BIP44 account with 2 UTXOs
-        let mut bip44_account = ManagedCoreAccount::new(
+        let mut bip44_account = ManagedCoreFundsAccount::new(
             ManagedAccountType::Standard {
                 index: 0,
                 standard_account_type: StandardAccountType::BIP44Account,
@@ -318,7 +318,7 @@ mod utxo_tests {
         managed_info.accounts.insert(bip44_account).unwrap();
 
         // Create BIP32 account with 1 UTXO
-        let mut bip32_account = ManagedCoreAccount::new(
+        let mut bip32_account = ManagedCoreFundsAccount::new(
             ManagedAccountType::Standard {
                 index: 0,
                 standard_account_type: StandardAccountType::BIP32Account,
@@ -344,7 +344,7 @@ mod utxo_tests {
         managed_info.accounts.insert(bip32_account).unwrap();
 
         // Create CoinJoin account with 2 UTXOs
-        let mut coinjoin_account = ManagedCoreAccount::new(
+        let mut coinjoin_account = ManagedCoreFundsAccount::new(
             ManagedAccountType::CoinJoin {
                 index: 0,
                 addresses: key_wallet::managed_account::address_pool::AddressPoolBuilder::default()
@@ -383,7 +383,7 @@ mod utxo_tests {
     fn test_managed_wallet_get_utxos() {
         use crate::managed_wallet::FFIManagedWalletInfo;
         use key_wallet::account::account_type::StandardAccountType;
-        use key_wallet::managed_account::ManagedCoreAccount;
+        use key_wallet::managed_account::ManagedCoreFundsAccount;
         use key_wallet::wallet::managed_wallet_info::ManagedWalletInfo;
         use key_wallet::Network;
 
@@ -396,7 +396,7 @@ mod utxo_tests {
         let mut managed_info = ManagedWalletInfo::new(Network::Testnet, [3u8; 32]);
 
         // Add a UTXO to Testnet account
-        let mut testnet_account = ManagedCoreAccount::new(
+        let mut testnet_account = ManagedCoreFundsAccount::new(
             ManagedAccountType::Standard {
                 index: 0,
                 standard_account_type: StandardAccountType::BIP44Account,
