+++ key-wallet/IMPLEMENTATION_SUMMARY.md
@@ -180,7 +180,7 @@ let account = wallet.create_account(
 )?;
 
 // Wrap in a managed account for mutable operations
-let mut managed = ManagedCoreAccount::from_account(&account);
+let mut managed = ManagedCoreFundsAccount::from_account(&account);
 
 // Get a receive address
 let address = managed.get_next_receive_address()?;
