+++ key-wallet/src/tests/spent_outpoints_tests.rs
@@ -5,7 +5,7 @@ use dashcore::{TxIn, Txid};
 
 use crate::account::{AccountType, StandardAccountType, TransactionRecord};
 use crate::managed_account::transaction_record::TransactionDirection;
-use crate::managed_account::ManagedCoreAccount;
+use crate::managed_account::ManagedCoreFundsAccount;
 use crate::transaction_checking::{TransactionContext, TransactionType};
 
 /// Create a transaction that spends the given outpoints.
@@ -54,14 +54,14 @@ fn record_from_tx(tx: &Transaction) -> TransactionRecord {
 
 #[test]
 fn fresh_account_has_empty_spent_outpoints() {
-    let account = ManagedCoreAccount::dummy_bip44();
+    let account = ManagedCoreFundsAccount::dummy_bip44();
     assert!(account.transactions.is_empty());
 
     let probe = OutPoint::new(Txid::from([0xAA; 32]), 0);
     // Accessing spent_outpoints on a fresh account should not panic or misbehave.
     // We verify indirectly via serde round-trip (spent_outpoints is private).
     let json = serde_json::to_string(&account).unwrap();
-    let deserialized: ManagedCoreAccount = serde_json::from_str(&json).unwrap();
+    let deserialized: ManagedCoreFundsAccount = serde_json::from_str(&json).unwrap();
     // No transactions, so spent_outpoints stays empty after round-trip.
     assert!(deserialized.transactions.is_empty());
     // Confirm the serialized form does not contain spent_outpoints.
@@ -71,7 +71,7 @@ fn fresh_account_has_empty_spent_outpoints() {
 
 #[test]
 fn serde_round_trip_rebuilds_spent_outpoints() {
-    let mut account = ManagedCoreAccount::dummy_bip44();
+    let mut account = ManagedCoreFundsAccount::dummy_bip44();
 
     let outpoint_a = OutPoint::new(Txid::from([0x01; 32]), 0);
     let outpoint_b = OutPoint::new(Txid::from([0x02; 32]), 1);
@@ -84,20 +84,20 @@ fn serde_round_trip_rebuilds_spent_outpoints() {
     assert!(!json.contains("spent_outpoints"));
 
     // Deserialize: spent_outpoints should be rebuilt from transactions
-    let deserialized: ManagedCoreAccount = serde_json::from_str(&json).unwrap();
+    let deserialized: ManagedCoreFundsAccount = serde_json::from_str(&json).unwrap();
     assert_eq!(deserialized.transactions.len(), 1);
 
     // Verify the rebuilt set by serializing again and comparing transactions
     // (spent_outpoints is private, so we test behavior through a second round-trip
     //  to confirm stability)
     let json2 = serde_json::to_string(&deserialized).unwrap();
-    let deserialized2: ManagedCoreAccount = serde_json::from_str(&json2).unwrap();
+    let deserialized2: ManagedCoreFundsAccount = serde_json::from_str(&json2).unwrap();
     assert_eq!(deserialized2.transactions.len(), 1);
 }
 
 #[test]
 fn receive_only_account_round_trips_correctly() {
-    let mut account = ManagedCoreAccount::dummy_bip44();
+    let mut account = ManagedCoreFundsAccount::dummy_bip44();
 
     // Add a receive-only transaction (coinbase-like, no real spent outpoints)
     let tx = receive_only_tx();
@@ -108,18 +108,18 @@ fn receive_only_account_round_trips_correctly() {
 
     // Round-trip should work without issues (no rebuild loop)
     let json = serde_json::to_string(&account).unwrap();
-    let deserialized: ManagedCoreAccount = serde_json::from_str(&json).unwrap();
+    let deserialized: ManagedCoreFundsAccount = serde_json::from_str(&json).unwrap();
     assert_eq!(deserialized.transactions.len(), 1);
 
     // A second round-trip should be stable
     let json2 = serde_json::to_string(&deserialized).unwrap();
-    let deserialized2: ManagedCoreAccount = serde_json::from_str(&json2).unwrap();
+    let deserialized2: ManagedCoreFundsAccount = serde_json::from_str(&json2).unwrap();
     assert_eq!(deserialized2.transactions.len(), 1);
 }
 
 #[test]
 fn multiple_transactions_all_inputs_tracked_after_round_trip() {
-    let mut account = ManagedCoreAccount::dummy_bip44();
+    let mut account = ManagedCoreFundsAccount::dummy_bip44();
 
     let outpoint_1 = OutPoint::new(Txid::from([0x10; 32]), 0);
     let outpoint_2 = OutPoint::new(Txid::from([0x20; 32]), 0);
@@ -132,7 +132,7 @@ fn multiple_transactions_all_inputs_tracked_after_round_trip() {
     account.transactions.insert(tx2.txid(), record_from_tx(&tx2));
 
     let json = serde_json::to_string(&account).unwrap();
-    let deserialized: ManagedCoreAccount = serde_json::from_str(&json).unwrap();
+    let deserialized: ManagedCoreFundsAccount = serde_json::from_str(&json).unwrap();
 
     // All three outpoints should be in the rebuilt spent set.
     // We verify by confirming the transaction inputs survived the round-trip.
