+++ key-wallet/src/wallet/managed_wallet_info/helpers.rs
@@ -2,24 +2,24 @@
 
 use super::ManagedWalletInfo;
 use crate::account::account_collection::PlatformPaymentAccountKey;
-use crate::account::ManagedCoreAccount;
+use crate::account::ManagedCoreFundsAccount;
 use crate::managed_account::managed_platform_account::ManagedPlatformAccount;
 
 impl ManagedWalletInfo {
     // BIP44 Account Helpers
 
     /// Get the first BIP44 managed account
-    pub fn first_bip44_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn first_bip44_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.bip44_managed_account_at_index(0)
     }
 
     /// Get the first BIP44 managed account (mutable)
-    pub fn first_bip44_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn first_bip44_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.bip44_managed_account_at_index_mut(0)
     }
 
     /// Get a BIP44 managed account at a specific index
-    pub fn bip44_managed_account_at_index(&self, index: u32) -> Option<&ManagedCoreAccount> {
+    pub fn bip44_managed_account_at_index(&self, index: u32) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.standard_bip44_accounts.get(&index)
     }
 
@@ -27,24 +27,24 @@ impl ManagedWalletInfo {
     pub fn bip44_managed_account_at_index_mut(
         &mut self,
         index: u32,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.standard_bip44_accounts.get_mut(&index)
     }
 
     // BIP32 Account Helpers
 
     /// Get the first BIP32 managed account
-    pub fn first_bip32_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn first_bip32_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.bip32_managed_account_at_index(0)
     }
 
     /// Get the first BIP32 managed account (mutable)
-    pub fn first_bip32_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn first_bip32_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.bip32_managed_account_at_index_mut(0)
     }
 
     /// Get a BIP32 managed account at a specific index
-    pub fn bip32_managed_account_at_index(&self, index: u32) -> Option<&ManagedCoreAccount> {
+    pub fn bip32_managed_account_at_index(&self, index: u32) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.standard_bip32_accounts.get(&index)
     }
 
@@ -52,24 +52,24 @@ impl ManagedWalletInfo {
     pub fn bip32_managed_account_at_index_mut(
         &mut self,
         index: u32,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.standard_bip32_accounts.get_mut(&index)
     }
 
     // CoinJoin Account Helpers
 
     /// Get the first CoinJoin managed account
-    pub fn first_coinjoin_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn first_coinjoin_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.coinjoin_managed_account_at_index(0)
     }
 
     /// Get the first CoinJoin managed account (mutable)
-    pub fn first_coinjoin_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn first_coinjoin_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.coinjoin_managed_account_at_index_mut(0)
     }
 
     /// Get a CoinJoin managed account at a specific index
-    pub fn coinjoin_managed_account_at_index(&self, index: u32) -> Option<&ManagedCoreAccount> {
+    pub fn coinjoin_managed_account_at_index(&self, index: u32) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.coinjoin_accounts.get(&index)
     }
 
@@ -77,19 +77,19 @@ impl ManagedWalletInfo {
     pub fn coinjoin_managed_account_at_index_mut(
         &mut self,
         index: u32,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.coinjoin_accounts.get_mut(&index)
     }
 
     // TopUp Account Helpers
 
     /// Get the first TopUp managed account
-    pub fn first_topup_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn first_topup_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.identity_topup.values().next()
     }
 
     /// Get the first TopUp managed account (mutable)
-    pub fn first_topup_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn first_topup_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.identity_topup.values_mut().next()
     }
 
@@ -97,7 +97,7 @@ impl ManagedWalletInfo {
     pub fn topup_managed_account_at_registration_index(
         &self,
         registration_index: u32,
-    ) -> Option<&ManagedCoreAccount> {
+    ) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.identity_topup.get(&registration_index)
     }
 
@@ -105,97 +105,97 @@ impl ManagedWalletInfo {
     pub fn topup_managed_account_at_registration_index_mut(
         &mut self,
         registration_index: u32,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.identity_topup.get_mut(&registration_index)
     }
 
     // Identity Registration Account Helper
 
     /// Get the identity registration managed account
-    pub fn identity_registration_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn identity_registration_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.identity_registration.as_ref()
     }
 
     /// Get the identity registration managed account (mutable)
-    pub fn identity_registration_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn identity_registration_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.identity_registration.as_mut()
     }
 
     // Identity TopUp Not Bound Account Helper
 
     /// Get the identity top-up not bound managed account
-    pub fn identity_topup_not_bound_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn identity_topup_not_bound_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.identity_topup_not_bound.as_ref()
     }
 
     /// Get the identity top-up not bound managed account (mutable)
     pub fn identity_topup_not_bound_managed_account_mut(
         &mut self,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.identity_topup_not_bound.as_mut()
     }
 
     // Identity Invitation Account Helper
 
     /// Get the identity invitation managed account
-    pub fn identity_invitation_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn identity_invitation_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.identity_invitation.as_ref()
     }
 
     /// Get the identity invitation managed account (mutable)
-    pub fn identity_invitation_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn identity_invitation_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.identity_invitation.as_mut()
     }
 
     // Provider Voting Keys Account Helper
 
     /// Get the provider voting keys managed account
-    pub fn provider_voting_keys_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn provider_voting_keys_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.provider_voting_keys.as_ref()
     }
 
     /// Get the provider voting keys managed account (mutable)
-    pub fn provider_voting_keys_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn provider_voting_keys_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.provider_voting_keys.as_mut()
     }
 
     // Provider Owner Keys Account Helper
 
     /// Get the provider owner keys managed account
-    pub fn provider_owner_keys_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn provider_owner_keys_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.provider_owner_keys.as_ref()
     }
 
     /// Get the provider owner keys managed account (mutable)
-    pub fn provider_owner_keys_managed_account_mut(&mut self) -> Option<&mut ManagedCoreAccount> {
+    pub fn provider_owner_keys_managed_account_mut(&mut self) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.provider_owner_keys.as_mut()
     }
 
     // Provider Operator Keys Account Helper
 
     /// Get the provider operator keys managed account
-    pub fn provider_operator_keys_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn provider_operator_keys_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.provider_operator_keys.as_ref()
     }
 
     /// Get the provider operator keys managed account (mutable)
     pub fn provider_operator_keys_managed_account_mut(
         &mut self,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.provider_operator_keys.as_mut()
     }
 
     // Provider Platform Keys Account Helper
 
     /// Get the provider platform keys managed account
-    pub fn provider_platform_keys_managed_account(&self) -> Option<&ManagedCoreAccount> {
+    pub fn provider_platform_keys_managed_account(&self) -> Option<&ManagedCoreFundsAccount> {
         self.accounts.provider_platform_keys.as_ref()
     }
 
     /// Get the provider platform keys managed account (mutable)
     pub fn provider_platform_keys_managed_account_mut(
         &mut self,
-    ) -> Option<&mut ManagedCoreAccount> {
+    ) -> Option<&mut ManagedCoreFundsAccount> {
         self.accounts.provider_platform_keys.as_mut()
     }
 
@@ -298,7 +298,7 @@ impl ManagedWalletInfo {
     }
 
     /// Get all accounts
-    pub fn all_managed_accounts(&self) -> Vec<&ManagedCoreAccount> {
+    pub fn all_managed_accounts(&self) -> Vec<&ManagedCoreFundsAccount> {
         self.accounts.all_accounts()
     }
 }
