+++ key-wallet/src/wallet/managed_wallet_info/managed_accounts.rs
@@ -7,7 +7,7 @@ use super::{managed_account_operations::ManagedAccountOperations, ManagedWalletI
 use crate::account::BLSAccount;
 #[cfg(feature = "eddsa")]
 use crate::account::EdDSAAccount;
-use crate::account::{Account, AccountType, ManagedCoreAccount};
+use crate::account::{Account, AccountType, ManagedCoreFundsAccount};
 use crate::bip32::ExtendedPubKey;
 use crate::error::{Error, Result};
 use crate::wallet::{Wallet, WalletType};
@@ -42,7 +42,7 @@ impl ManagedAccountOperations for ManagedWalletInfo {
         })?;
 
         // Create the ManagedAccount from the Account
-        let managed_account = ManagedCoreAccount::from_account(account);
+        let managed_account = ManagedCoreFundsAccount::from_account(account);
 
         // Check if managed account already exists
         if self.accounts.contains_managed_account_type(managed_account.managed_type()) {
@@ -117,7 +117,7 @@ impl ManagedAccountOperations for ManagedWalletInfo {
         let account = Account::new(None, account_type, account_xpub, self.network)?;
 
         // Create the ManagedAccount from the Account
-        let managed_account = ManagedCoreAccount::from_account(&account);
+        let managed_account = ManagedCoreFundsAccount::from_account(&account);
 
         // Check if managed account already exists
         if self.accounts.contains_managed_account_type(managed_account.managed_type()) {
@@ -162,7 +162,7 @@ impl ManagedAccountOperations for ManagedWalletInfo {
         })?;
 
         // Create the ManagedAccount from the BLS Account
-        let managed_account = ManagedCoreAccount::from_bls_account(bls_account);
+        let managed_account = ManagedCoreFundsAccount::from_bls_account(bls_account);
 
         // Check if managed account already exists
         if self.accounts.contains_managed_account_type(managed_account.managed_type()) {
@@ -234,7 +234,7 @@ impl ManagedAccountOperations for ManagedWalletInfo {
             BLSAccount::from_public_key_bytes(None, account_type, bls_public_key, self.network)?;
 
         // Create the ManagedAccount from the BLS Account
-        let managed_account = ManagedCoreAccount::from_bls_account(&bls_account);
+        let managed_account = ManagedCoreFundsAccount::from_bls_account(&bls_account);
 
         // Check if managed account already exists
         if self.accounts.contains_managed_account_type(managed_account.managed_type()) {
@@ -280,7 +280,7 @@ impl ManagedAccountOperations for ManagedWalletInfo {
             })?;
 
         // Create the ManagedAccount from the EdDSA Account
-        let managed_account = ManagedCoreAccount::from_eddsa_account(eddsa_account);
+        let managed_account = ManagedCoreFundsAccount::from_eddsa_account(eddsa_account);
 
         // Check if managed account already exists
         if self.accounts.contains_managed_account_type(managed_account.managed_type()) {
@@ -356,7 +356,7 @@ impl ManagedAccountOperations for ManagedWalletInfo {
         )?;
 
         // Create the ManagedAccount from the EdDSA Account
-        let managed_account = ManagedCoreAccount::from_eddsa_account(&eddsa_account);
+        let managed_account = ManagedCoreFundsAccount::from_eddsa_account(&eddsa_account);
 
         // Check if managed account already exists
         if self.accounts.contains_managed_account_type(managed_account.managed_type()) {
